requirements.txt
# Core dependencies
llama-cpp-python>=0.2.0
opencv-python>=4.8.0
mss>=9.0.1
numpy>=1.24.0

# Deep learning (PPO)
torch>=2.0.0

# Input simulation
pyautogui>=0.9.54

# Configuration
pyyaml>=6.0

# Imaging utilities
Pillow>=10.0.0
